+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Throw an error on unknown gg one sub commands, also if option -h is given
+
 ## 12.3.0 - 2026-08-01
 
 ### Changed